fix: changing the positions of share feature on videos and podcasts episode (#3540)
* fix: changing the positions of share feature on videos and podcasts episode



---------
