+++ frontends/main/src/page-components/ResourceCard/ResourceCard.tsx
@@ -8,16 +8,11 @@ import {
 } from "../Dialogs/AddToListDialog"
 import { useResourceDrawerHref } from "../LearningResourceDrawer/useResourceDrawerHref"
 import { useUserMe } from "api/hooks/user"
-import { LearningResource, PodcastEpisodeResource, ResourceTypeEnum } from "api"
+import { LearningResource } from "api"
 import { SignupPopover } from "../SignupPopover/SignupPopover"
 import { useIsUserListMember } from "api/hooks/userLists"
 import { useLearningResourceDetailSetCache } from "api/hooks/learningResources"
 import { useIsLearningPathMember } from "api/hooks/learningPaths"
-import ShareDialog from "@/app-pages/VideoPlaylistCollectionPage/ShareDialog"
-import { env } from "@/env"
-import { podcastEpisodePageView } from "@/common/urls"
-
-const NEXT_PUBLIC_ORIGIN = env("NEXT_PUBLIC_ORIGIN")
 
 export const useResourceCard = (resource?: LearningResource | null) => {
   const getDrawerHref = useResourceDrawerHref()
@@ -26,7 +21,6 @@ export const useResourceCard = (resource?: LearningResource | null) => {
   const { data: inLearningPath } = useIsLearningPathMember(resource?.id)
 
   const [anchorEl, setAnchorEl] = useState<HTMLElement | null>(null)
-  const [shareAnchorEl, setShareAnchorEl] = useState<HTMLElement | null>(null)
 
   const handleClosePopover = useCallback(() => {
     setAnchorEl(null)
@@ -57,11 +51,6 @@ export const useResourceCard = (resource?: LearningResource | null) => {
       }
     }, [user])
 
-  const handleShareClick: LearningResourceCardProps["onShareClick"] =
-    useCallback((event: React.MouseEvent<HTMLButtonElement>) => {
-      setShareAnchorEl(event.currentTarget)
-    }, [])
-
   const onClick = useLearningResourceDetailSetCache(resource)
 
   return {
@@ -71,9 +60,6 @@ export const useResourceCard = (resource?: LearningResource | null) => {
     handleClosePopover,
     handleAddToLearningPathClick,
     handleAddToUserListClick,
-    handleShareClick,
-    shareAnchorEl,
-    setShareAnchorEl,
     inUserList,
     inLearningPath,
   }
@@ -92,7 +78,7 @@ const subheadingMap: Record<HeadingElement, number> = {
 
 type ResourceCardProps = Omit<
   LearningResourceCardProps,
-  "href" | "onAddToLearningPathClick" | "onAddToUserListClick" | "onShareClick"
+  "href" | "onAddToLearningPathClick" | "onAddToUserListClick"
 > & {
   headingLevel?: number
   parentHeadingEl?: HeadingElement
@@ -118,9 +104,6 @@ const ResourceCard: React.FC<ResourceCardProps> = ({
     handleClosePopover,
     handleAddToLearningPathClick,
     handleAddToUserListClick,
-    handleShareClick,
-    shareAnchorEl,
-    setShareAnchorEl,
     inUserList,
     inLearningPath,
     onClick,
@@ -135,16 +118,6 @@ const ResourceCard: React.FC<ResourceCardProps> = ({
 
   const headingLevel = parentHeadingEl ? subheadingMap[parentHeadingEl] : 6
 
-  const isPodcastEpisode =
-    resource?.resource_type === ResourceTypeEnum.PodcastEpisode
-  const podcastId = isPodcastEpisode
-    ? resource?.podcast_episode?.podcasts?.[0]
-    : undefined
-  const sharePageUrl =
-    isPodcastEpisode && podcastId !== undefined
-      ? `${NEXT_PUBLIC_ORIGIN}${podcastEpisodePageView(String(resource!.id), String(podcastId), resource?.title)}`
-      : ""
-
   return (
     <>
       <LearningResourceCard
@@ -153,24 +126,12 @@ const ResourceCard: React.FC<ResourceCardProps> = ({
         href={resource ? getDrawerHref(resource.id) : undefined}
         onAddToLearningPathClick={handleAddToLearningPathClick}
         onAddToUserListClick={handleAddToUserListClick}
-        onShareClick={
-          isPodcastEpisode && podcastId !== undefined ? handleShareClick : null
-        }
         inUserList={inUserList}
         inLearningPath={inLearningPath}
         headingLevel={headingLevel}
         {...others}
       />
       <SignupPopover anchorEl={anchorEl} onClose={handleClosePopover} />
-      {isPodcastEpisode && podcastId !== undefined && (
-        <ShareDialog
-          open={Boolean(shareAnchorEl)}
-          onClose={() => setShareAnchorEl(null)}
-          resource={resource as PodcastEpisodeResource}
-          pageUrl={sharePageUrl}
-          title={resource?.title ?? ""}
-        />
-      )}
     </>
   )
 }
